backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic-settings==2.8.1
redis==5.2.1
psycopg[binary]==3.2.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
langchain==0.3.23
langchain-openai==0.3.12
langchain-google-genai==2.0.10
langchain-groq==0.3.2
bcrypt==4.0.1

